# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httputil.Error400Response:
    properties:
      error:
        example: invalid request body
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error401Response:
    properties:
      error:
        example: unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  httputil.Error404Response:
    properties:
      error:
        example: resource not found
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error500Response:
    properties:
      error:
        example: internal server error
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.MessageResponse:
    properties:
      data:
        example: ok
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  modifier.CreateModifierGroupInput:
    properties:
      is_required:
        type: boolean
      max_select:
        type: integer
      name:
        type: string
      options:
        items:
          properties:
            extra_price:
              type: number
            name:
              type: string
          type: object
        type: array
    type: object
  modifier.ModifierGroup:
    properties:
      business_id:
        type: string
      id:
        type: string
      is_required:
        type: boolean
      max_select:
        type: integer
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/modifier.ModifierOption'
        type: array
    type: object
  modifier.ModifierOption:
    properties:
      extra_price:
        type: number
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  modifier.UpdateModifierGroupInput:
    properties:
      is_required:
        type: boolean
      max_select:
        type: integer
      name:
        type: string
    type: object
  order.AddOrderItemInput:
    properties:
      menu_item_id:
        type: string
      modifier_option_ids:
        items:
          type: string
        type: array
      notes:
        type: string
      quantity:
        type: integer
    type: object
  order.CreateOrderInput:
    properties:
      branch_id:
        type: string
      customer_name:
        type: string
      table_id:
        type: string
      type:
        type: string
    type: object
  order.Order:
    properties:
      branch_id:
        type: string
      business_id:
        type: string
      cashier_id:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order.OrderItem'
        type: array
      service_charge_amount:
        type: number
      status:
        type: string
      subtotal:
        type: number
      table_id:
        type: string
      tax_amount:
        type: number
      total:
        type: number
      type:
        type: string
    type: object
  order.OrderItem:
    properties:
      id:
        type: string
      menu_item_id:
        type: string
      modifiers:
        items:
          $ref: '#/definitions/order.OrderItemModifier'
        type: array
      notes:
        type: string
      order_id:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
      unit_price:
        type: number
    type: object
  order.OrderItemModifier:
    properties:
      extra_price:
        type: number
      id:
        type: string
      modifier_option_id:
        type: string
      order_item_id:
        type: string
    type: object
  order.UpdateOrderInput:
    properties:
      customer_name:
        type: string
      status:
        type: string
    type: object
  order.UpdateOrderItemInput:
    properties:
      modifier_option_ids:
        items:
          type: string
        type: array
      notes:
        type: string
      quantity:
        type: integer
    type: object
  payment.PayOrderInput:
    properties:
      amount_paid:
        type: number
      method:
        type: string
    type: object
  payment.Payment:
    properties:
      amount_paid:
        type: number
      change_amount:
        type: number
      id:
        type: string
      method:
        type: string
      order_id:
        type: string
      paid_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant POS SaaS backend
  title: Bayar.in API
  version: "1.0"
paths:
...
  /modifiers:
    get:
      description: Returns all modifier groups (with their options) for the authenticated
        owner's business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/modifier.ModifierGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: List modifier groups
      tags:
      - Modifiers
    post:
      consumes:
      - application/json
      description: Create a modifier group with options in a single transaction
      parameters:
      - description: Modifier group payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/modifier.CreateModifierGroupInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/modifier.ModifierGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Create modifier group
      tags:
      - Modifiers
  /modifiers/{id}:
    patch:
      consumes:
      - application/json
      description: Partially update a modifier group's name, required flag, or max
        selections
      parameters:
      - description: Modifier group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/modifier.UpdateModifierGroupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modifier.ModifierGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update modifier group
      tags:
      - Modifiers
  /orders:
    get:
      description: 'Returns orders for the business. Cashiers see only their branch''s
        orders. Filter by status: open|paid|cancelled.'
      parameters:
      - description: Filter by status (open, paid, cancelled)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order. Type must be "dine_in" or "takeaway". table_id
        is optional.
      parameters:
      - description: Order payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/order.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - Orders
  /orders/{id}:
    get:
      description: Returns a single order with all its items and applied modifiers
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.Order'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error404Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
    patch:
      consumes:
      - application/json
      description: Update order status or customer name. Cannot update paid or cancelled
        orders.
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update order
      tags:
      - Orders
  /orders/{id}/items:
    post:
      consumes:
      - application/json
      description: Add a menu item to an open order with optional modifier options.
        Validates modifiers belong to the menu item.
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Order item payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/order.AddOrderItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/order.OrderItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error404Response'
      security:
      - BearerAuth: []
      summary: Add item to order
      tags:
      - orders
  /orders/{id}/items/{item_id}:
    delete:
      description: Remove an item from an open order. Recalculates order totals.
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Order item UUID
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Delete order item
      tags:
      - Order Items
    patch:
      consumes:
      - application/json
      description: Update quantity or notes of an existing order item. Only works
        on open orders.
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Order item UUID
        in: path
        name: item_id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update order item
      tags:
      - Order Items
  /orders/{id}/pay:
    post:
      consumes:
      - application/json
      description: Process payment for an open order. Method must be "cash", "qris",
        or "transfer". amount_paid must be >= order total. Atomically marks order
        as paid and releases table if dine_in.
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Payment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payment.PayOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payment.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Pay order
      tags:
      - Payment
...
securityDefinitions:
  BearerAuth:
    description: 'Paste your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"